parse ledger's xml format
I initially wanted to use streams for API calls to xml parsing, but
having strings in memory and setting up an adhoc stream on it did not
produce the results I wanted.

For some reason with a flowing stream coming from a string, expat would
not emit any event.

For this reason I'll simply feed the string directly to expat.

We parse transactions from xml:

- transaction dates
- collect payees
- collect postings

Quantities are stored in bignums, since we're talking money